+++ web/components/Hello.tsx
@@ -1,11 +1,16 @@
 import { FC } from 'react';
+import { Button } from '../ui/uikit/Button';
 
 interface HelloProps {
   name: string;
 }
 
 export const Hello: FC<HelloProps> = ({ name }) => {
-  return <div>Привет, {name}!</div>;
+  return (
+    <div>
+      Привет, {name}!<Button style={{ marginLeft: '0.5rem' }}>OK</Button>
+    </div>
+  );
 };
 
 export default Hello;